Load the tower dataset and create a plot.

plot.hdf5 loads the tower data from an HDF5 file generated by ED and will plot the values against one another. The default is for the given variable to be plotted against time.

plot.netcdf(datafile, yvar, xvar = "time", width = 800, height = 600,
  filename = NULL, year = NULL)

Arguments

datafile
the specific datafile to use.

yvar
the variable to plot along the y-axis.

xvar
the variable to plot along the x-axis, by default time is used.

filename
is the name of the file name that is geneated, this can be null to use existing device, otherwise it will try and

year
the year this data is for (only used in the title).

the
width of the image generated, default is 800 pixels.

the
height of the image generated, default is 600 pixels.